`Item` struct data types rework.

+++ src/item/Item.go
@@ -1,9 +1,9 @@
 package item
 
 type Item struct {
-	Id               uint8            `json:"id"`
+	Id               int              `json:"id"`
 	Name             string           `json:"name"`
-	PreparationTime  uint             `json:"preparation-time"`
-	Complexity       uint8            `json:"complexity"`
+	PreparationTime  int              `json:"preparation-time"`
+	Complexity       int              `json:"complexity"`
 	CookingApparatus CookingApparatus `json:"cooking-apparatus"`
 }
